+++ src/ApiClient.php
@@ -40,7 +40,7 @@ final class ApiClient
     use Endpoints\Customers;
     use Endpoints\Addresses;
 
-    public const VERSION = '0.7.0';
+    public const VERSION = '0.8.0';
 
     private const SANDBOX_URL = 'https://sandbox.v-shop.cloud/';
     private const SANDBOX_CLIENT_ID = 'test+api@sandbox.v-shop.cloud';